checkpoint a refactor of maxwell parser run-loop
